Introduce the use of application contexts
This commit introduces Flask to help manage the problem of global
state.

Particularly this adds:
- An application factory function for generating app objects
- Implicit creation of an application object on server start
- A pytest fixture for an app

At the beginning of a request, an app context is pushed manually. This
is typically taken care of within the Flask framework when using their
request mechanisms, but we're only using Flask for the app context.

Finally, there were a couple of tests that required a server be
started that weren't using the fixture.